changed to gcm and added message signing

+++ client/storage/sessionManagement.py
@@ -1,4 +1,3 @@
-import os
 import keyring
 import sys
 from keyrings.alt.file import PlaintextKeyring
@@ -20,7 +19,7 @@ def setCredentials(userId, privateKey, publicKey, password, iv):
   keyring.set_password(serviceP, username, password)
   keyring.set_password(serviceU, username, userId)
   keyring.set_password(serviceIV, username, iv)
-  print('successfully added credentials to keyring')
+  print('Successfully added credentials to keyring')
 
 def initiateConvo(userId, publicKey):
   keyring.set_password(convoUserId, username, userId)
@@ -89,10 +88,15 @@ def checkIfLoggedIn():
     sys.exit(1)
 
 def removeCredentials():
+  if keyring.get_password(serviceP, username) is None:
+    print('Not Currently Logged In')
+    sys.exit(1)
   keyring.delete_password(servicePrK, username)
   keyring.delete_password(servicePbK, username)
   keyring.delete_password(serviceP, username)
   keyring.delete_password(serviceU, username)
   keyring.delete_password(serviceIV, username)
-  keyring.delete_password(convoUserId, username)
-  keyring.delete_password(convoPublicKey, username)
+  if keyring.get_password(convoUserId, username) is not None:
+    keyring.delete_password(convoUserId, username)
+    keyring.delete_password(convoPublicKey, username)
+
